% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{as_rule_df}
\alias{as_rule_df}
\title{Helper to class a Snort rules data frame properly}
\usage{
as_rule_df(rules)
}
\arguments{
\item{rules}{a Snort rules data frame read in with [read_rules()].}
}
\description{
Helper to class a Snort rules data frame properly
}
