% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-rules.R
\name{download_community_rules}
\alias{download_community_rules}
\title{Download Snort community rules}
\usage{
download_community_rules(outpath = NULL)
}
\arguments{
\item{outpath}{output path+file. If \code{NULL} then a date-stamped file will be
created in the current working directory}
}
\description{
Download Snort community rules
}
