% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-rules.R
\name{download_subscription_rules}
\alias{download_subscription_rules}
\title{Download Snort subscription rules}
\usage{
download_subscription_rules(version = NULL, outpath = NULL,
  overwrite = FALSE, oinkcode = Sys.getenv("OINKCODE"))
}
\arguments{
\item{version}{if \code{NULL} then the latest version will be used, otherwise provide
the version number (with or without dots)}

\item{outpath}{output path+file. If \code{NULL} then a date-stamped file will be
created in the current working directory}

\item{overwrite}{if \code{TRUE} and the \code{outpath} file exists then it will be overwritten.
Defaults to \code{TRUE}}

\item{oinkcode}{your Snort oinkcode (ideally stored in \code{OINKCODE} env var)}
}
\description{
Download Snort subscription rules
}
