% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-range.R
\name{expand_port_ranges}
\alias{expand_port_ranges}
\title{Expand a Snort port description}
\usage{
expand_port_ranges(x)
}
\arguments{
\item{x}{character vector of individual port strings}
}
\value{
list
}
\description{
Snort rule port descriptions can be anything from a single number, to \code{any},
to a \code{$} variable specifier to a full or partial range. This function will
identify full or partial ranges and expand then, leaving other port descriptors
untouched.
}
\examples{
expand_port_ranges(c("25", "$HTTP_PORTS", "1024:", ":1024", "1:1024", "any"))
}
