% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porc-package.R
\docType{package}
\name{porc}
\alias{porc}
\alias{porc-package}
\title{Tools to Work with 'Snort' Rules, Logs and Data}
\description{
'Snort' is an open source intrusion prevention system capable of
real-time traffic analysis and packet logging. Tools are provided to
work with 'Snort' rulesets, logs and other data associated with the platform.
More information on 'Snort' can be found at <https://www.snort.org/>.
}
\author{
Bob Rudis (bob@rud.is)
}
