% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendedv2.R
\name{read_extended}
\alias{read_extended}
\title{Read a Snort "extended v2" log into a data frame}
\usage{
read_extended(path)
}
\arguments{
\item{path}{path to log file}
}
\value{
data frame
}
\description{
Read a Snort "extended v2" log into a data frame
}
\examples{
read_extended(system.file("extdata", "multi-record-event-x2.log", package="porc"))
}
