% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-rules.R
\name{read_rules}
\alias{read_rules}
\title{Parse in a file of snort rules into a data frame}
\usage{
read_rules(path)
}
\arguments{
\item{path}{path to snort rules file}
}
\value{
data frame
}
\description{
The individual components of a Snort rule are parsed and put into a tidy
data frame format. The rule options are also parsed and the resultant
\code{options} columns is a data frame with option name and value. The value
for options that do not have values is \code{NA_character}.\cr
\cr
There is an \code{id} column which is just an index of the rule position in the
file. An extra \code{commented} field exists to identify rules that are in the
file but commented out. This makes it easy to filter on in-use rules.\cr
\cr
The \code{options} column can be unnested with \code{tidyr::unnest()}.
}
\examples{
rules <- read_rules(
  system.file("extdata", "emerging-telnet.rules", package="porc")
)
}
