% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule-vars.R
\name{rule_vars}
\alias{rule_vars}
\title{Extract all the `$`-named variables from Snort rules}
\usage{
rule_vars(rules)
}
\arguments{
\item{rules}{a Snort rules data frame read in with [read_rules()].}
}
\description{
Extract all the `$`-named variables from Snort rules
}
\examples{
rules <- read_rules(
  system.file("extdata", "emerging-telnet.rules", package="porc")
)

rule_vars(rules)
}
